Modification de la classe Zoom
