--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "Assets/Plugins/DOTween"]
-	path = Assets/Plugins/DOTween
-	url = https://github.com/Demigiant/dotween.git
