+++ Assets/Scripts/Menu/Button/ParticularButton/ButtonQuitSettingsAction.cs
@@ -5,6 +5,7 @@ public class ButtonQuitSettingsAction : ButtonAction
 {
 
     [SerializeField] private GameObject settingsButton;
+    [SerializeField] private ButtonQuitSettingsAction settingButtonAction;
 
     private Vector3 initialSettingsButtonPos;
