--- Assets/Scripts/Utility.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 92b408dd433e5f441948fd699f717590
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
