--- Assets/Scripts/Utility/Easing.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9dd0736693e440246886d870e6dc515e
